// backend/src/api/mil/enums/competency-level.enum.ts
export enum CompetencyLevel {
  BEGINNING = 'BEGINNING',
  DEVELOPING = 'DEVELOPING',
  PROFICIENT = 'PROFICIENT',
  ADVANCED = 'ADVANCED',
}

export enum CompetencyEvidenceSource {
  GUIDED_INVESTIGATION = 'GUIDED_INVESTIGATION',
  BASELINE_ASSESSMENT = 'BASELINE_ASSESSMENT',
  FOLLOW_UP_ASSESSMENT = 'FOLLOW_UP_ASSESSMENT',
  QUIZ = 'QUIZ',
  CHALLENGE = 'CHALLENGE',
}